Delete command updated

+++ src/Commands/CrudGeneratorDelete.php
@@ -38,6 +38,10 @@ class CrudGeneratorDelete extends Command
      *
      * @return void
      */
+    protected $api_version;
+    protected $api_auth;
+    protected $apiRouteMiddleware;
+    
     protected $version;
     protected $crudType;
     protected $name;
@@ -55,7 +59,11 @@ class CrudGeneratorDelete extends Command
     public function __construct(Filesystem $finder)
     {
         parent::__construct();
-        $this->version = 'v1';
+        $this->api_version = config('crudgenerator.api_version') ?? '1';
+        $this->api_auth = config('crudgenerator.api_auth') ?? 'sanctum';
+        $this->apiRouteMiddleware = $this->api_auth === 'sanctum' ? 'sanctum' : 'api';
+
+        $this->version = $this->api_version ? "v{$this->api_version}" : ''; //it may change but command should not change. conmmand should not contain v1 or etc
         $this->crudType = 'normal';
         $this->finder = $finder;
     }
@@ -155,27 +163,39 @@ class CrudGeneratorDelete extends Command
     protected function modelDelete()
     {
         //Version Check Code
-        $modelFolder = app()->version() < 8 ? '' : '/Models'; //laravel 8 uses Models folder
+        $modelFolder = intval(app()->version()) < 8 ? '' : '/Models'; //laravel 8 uses Models folder
         $path = app_path("$modelFolder/{$this->modelPascal}.php");
         
-        $this->validatePath($path);
-        $this->finder->delete($path);
+        $validated = $this->validatePath($path);
+
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('Model deleted successfully!');
+        }
     }
 
     protected function controllerDelete()
     {
         $path = app_path("/Http/Controllers{$this->adminFolder}/{$this->modelPascal}Controller.php");
 
-        $this->validatePath($path);
-        $this->finder->delete($path);
+        $validated = $this->validatePath($path);
+
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('Controller deleted successfully!');
+        }
     }
 
     protected function requestDelete()
     {
         $path = app_path("/Http/Requests/{$this->modelPascal}Request.php");
 
-        $this->validatePath($path);
-        $this->finder->delete($path);
+        $validated = $this->validatePath($path);
+
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('Request deleted successfully!');
+        }
     }
 
     protected function viewDelete()
@@ -185,8 +205,12 @@ class CrudGeneratorDelete extends Command
         $modellower = strtolower($this->modelSnake);
         $path = resource_path("views{$modelSnakeParent}/{$modellower}"); //view/backend or admin folder: get it from config
 
-        $this->validatePath($path);
-        $this->finder->deleteDirectory($path);
+        $validated = $this->validatePath($path);
+
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('views deleted successfully!');
+        }
     }
 
     protected function validatePath($path)
@@ -194,9 +218,9 @@ class CrudGeneratorDelete extends Command
         $this->protectCorePath($path);
         if ($this->finder->exists($path) === false) {
             $this->error("This $path does not exist!");
-            return;
+            return false;
         }
-        return;
+        return true;
     }
 
     protected function protectCorePath($path)
@@ -221,15 +245,27 @@ class CrudGeneratorDelete extends Command
     {
         $path = app_path("/Http/Resources/{$this->modelPascal}Resource.php");
 
-        $this->validatePath($path);
-        $this->finder->delete($path);
+        $validated = $this->validatePath($path);
+        
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('Api Resource deleted successfully!');
+        }
     }
 
     protected function apiControllerDelete()
     {
-        $path = app_path("/Http/Controllers/Api{$this->adminFolder}/{$this->modelPascal}Controller.php");
+        $apiVersion = ucfirst($this->version);
+        $apiVersion = ! empty($apiVersion) ? '/' . $apiVersion : '';
 
-        $this->validatePath($path);
-        $this->finder->delete($path);
+        $path = app_path("/Http/Controllers/Api{$apiVersion}{$this->adminFolder}/{$this->modelPascal}Controller.php");
+
+        $validated = $this->validatePath($path);
+
+        if ( $validated ) {
+            $this->finder->delete($path);
+            $this->info('Api Controller deleted successfully!');
+        }
+
     }
 }
